requirements.txt
pyOpenSSL
certifi
